Reorganize CLI commands into separate modules
